This copy of lamellar-benchmarks is a near-exact facsimile of the one which was used for scaling benchmarks on Sept 26, 2023. It is adapted from a commit on Dec 6, 2023, but all relevant files were individually checked and verified; they have no meaningful differences from the Sept 26, 2023 version:
- permute_matrix.rs
- toposort_atomic_dart.rs
- transpose.rs
- transpose_particles.rs
- transpose_particles_max.rs

Minimal changes have been made to the Dec 6, 2023 commit in order to for
archival purposes:

- binary files were updated to include a print statement on the last line,
  which helps automatic extraction of execution times
- the lamellar dependency in Cargo.toml was updated to
  ```
  lamellar =    {   git = "https://github.com/pnnl/lamellar-runtime",  rev = "b0489af", features = ["enable-rofi"]    }
  ```
  in order to fix the exact version which was used during benchmarking
- .gitignore was set to ignore .json files

**See also Dependencies, below.**

This repo contains at least one, but often more, binaries for a each sparse
matrix application. The specific files used for the reported benchmarks were

- permute_matrix.rs
- toposort_atomic_dart.rs
  - uses a faster algorithm to generate random matrices
- transpose_particles_max.rs
  - this had better performance than transpose.rs and transpose_particles.rs

We determined that the version of `lamellar-runtime` used for the benchmarks
was (possibly)
  - b0489af (short form)
  - b0489af041aca73a8b9f49de2c6b7e857689333b (long form)

The version of `rofi-sys` used by that copy of `lamellar-runtime` was specified
as
```
rofisys = { version ="0.2", optional = true }
```
in `lamellar-runtime/Cargo.toml`.

**However: do note that we encountered what appear to be some deadlocks** when
runninng this code. No deadlocks were reported in the original benchmarks,
so it is possible that an earlier version of lamellar-runtime was used at the
time. (Indeed, versiom `b0489af` was not the most recent commit to dev on
lamellar-runtime when the benchmarks originally ran; we inferred that the
benchmarks must have used an earlier version because there were many deadlocks
when we went back and tried using the version of dev that was current at
the time).